+++ composeunstyled-scrollbars/src/commonTest/kotlin/ScrollBars.test.kt
@@ -53,15 +53,15 @@ class ScrollBarsTest {
   fun always_visible_keeps_the_thumb_visible() = runComposeUiTest {
     setContent {
       val scrollState = rememberScrollState()
-      val scrollAreaState = rememberScrollAreaState(scrollState)
-      ScrollArea(state = scrollAreaState) {
+      val scrollbarState = rememberScrollbarState(scrollState)
+      ScrollArea(state = scrollbarState) {
         Column(modifier = Modifier.height(200.dp).verticalScroll(scrollState)) {
           repeat(100) { index ->
             BasicText("Item $index")
           }
         }
         UnstyledVerticalScrollbar(
-          scrollAreaState = scrollAreaState,
+          scrollbarState = scrollbarState,
           modifier = Modifier.testTag("scrollbar"),
         ) {
           Thumb(
@@ -81,15 +81,15 @@ class ScrollBarsTest {
   fun hide_while_idle_starts_with_thumb_out_of_composition() = runComposeUiTest {
     setContent {
       val scrollState = rememberScrollState()
-      val scrollAreaState = rememberScrollAreaState(scrollState)
-      ScrollArea(state = scrollAreaState) {
+      val scrollbarState = rememberScrollbarState(scrollState)
+      ScrollArea(state = scrollbarState) {
         Column(modifier = Modifier.height(200.dp).verticalScroll(scrollState)) {
           repeat(100) { index ->
             BasicText("Item $index")
           }
         }
         UnstyledVerticalScrollbar(
-          scrollAreaState = scrollAreaState,
+          scrollbarState = scrollbarState,
           modifier = Modifier.testTag("scrollbar"),
         ) {
           Thumb(
@@ -111,15 +111,15 @@ class ScrollBarsTest {
   fun hide_while_idle_shows_thumb_when_list_is_scrolled() = runComposeUiTest {
     setContent {
       val scrollState = rememberScrollState()
-      val scrollAreaState = rememberScrollAreaState(scrollState)
-      ScrollArea(state = scrollAreaState) {
+      val scrollbarState = rememberScrollbarState(scrollState)
+      ScrollArea(state = scrollbarState) {
         Column(modifier = Modifier.height(200.dp).verticalScroll(scrollState).testTag("list")) {
           repeat(100) { index ->
             BasicText("Item $index")
           }
         }
         UnstyledVerticalScrollbar(
-          scrollAreaState = scrollAreaState,
+          scrollbarState = scrollbarState,
           modifier = Modifier.testTag("scrollbar"),
         ) {
           Thumb(
@@ -145,15 +145,15 @@ class ScrollBarsTest {
     // TODO this checks for drags, not scrolls
     setContent {
       val scrollState = rememberScrollState()
-      val scrollAreaState = rememberScrollAreaState(scrollState)
-      ScrollArea(state = scrollAreaState) {
+      val scrollbarState = rememberScrollbarState(scrollState)
+      ScrollArea(state = scrollbarState) {
         Column(modifier = Modifier.height(200.dp).verticalScroll(scrollState).testTag("list")) {
           repeat(100) { index ->
             BasicText("Item $index")
           }
         }
         UnstyledVerticalScrollbar(
-          scrollAreaState = scrollAreaState,
+          scrollbarState = scrollbarState,
           modifier = Modifier.testTag("scrollbar")
             .align(Alignment.CenterEnd),
         ) {
@@ -190,8 +190,8 @@ class ScrollBarsTest {
   fun hide_while_idle_does_not_hide_thumb_while_dragging() = runComposeUiTest {
     setContent {
       val scrollState = rememberScrollState()
-      val scrollAreaState = rememberScrollAreaState(scrollState)
-      ScrollArea(state = scrollAreaState) {
+      val scrollbarState = rememberScrollbarState(scrollState)
+      ScrollArea(state = scrollbarState) {
         Column(
           modifier = Modifier
             .height(200.dp)
@@ -201,7 +201,7 @@ class ScrollBarsTest {
           repeat(100) { BasicText("Item $it") }
         }
         UnstyledVerticalScrollbar(
-          scrollAreaState = scrollAreaState,
+          scrollbarState = scrollbarState,
           modifier = Modifier
             .testTag("scrollbar")
             .align(Alignment.CenterEnd),
@@ -259,15 +259,15 @@ class ScrollBarsTest {
   fun hide_while_idle_does_not_hide_thumb_while_dragging_the_thumb() = runComposeUiTest {
     setContent {
       val scrollState = rememberScrollState()
-      val scrollAreaState = rememberScrollAreaState(scrollState)
-      ScrollArea(state = scrollAreaState) {
+      val scrollbarState = rememberScrollbarState(scrollState)
+      ScrollArea(state = scrollbarState) {
         Column(modifier = Modifier.height(200.dp).verticalScroll(scrollState).testTag("list")) {
           repeat(100) { index ->
             BasicText("Item $index")
           }
         }
         UnstyledVerticalScrollbar(
-          scrollAreaState = scrollAreaState,
+          scrollbarState = scrollbarState,
           modifier = Modifier.testTag("scrollbar"),
         ) {
           Thumb(
@@ -313,15 +313,15 @@ class ScrollBarsTest {
     runComposeUiTest {
       setContent {
         val scrollState = rememberScrollState()
-        val scrollAreaState = rememberScrollAreaState(scrollState)
-        ScrollArea(state = scrollAreaState) {
+        val scrollbarState = rememberScrollbarState(scrollState)
+        ScrollArea(state = scrollbarState) {
           Column(modifier = Modifier.height(200.dp).verticalScroll(scrollState).testTag("list")) {
             repeat(100) { index ->
               BasicText("Item $index")
             }
           }
           UnstyledVerticalScrollbar(
-            scrollAreaState = scrollAreaState,
+            scrollbarState = scrollbarState,
             modifier = Modifier.testTag("track"),
           ) {
             Thumb(
@@ -373,15 +373,15 @@ class ScrollBarsTest {
   fun vertical_scrollbar_uses_thumb_modifiers_for_minimum_size_and_thickness() = runComposeUiTest {
     setContent {
       val lazyListState = rememberLazyListState()
-      val scrollAreaState = rememberScrollAreaState(lazyListState)
-      ScrollArea(state = scrollAreaState) {
+      val scrollbarState = rememberScrollbarState(lazyListState)
+      ScrollArea(state = scrollbarState) {
         LazyColumn(modifier = Modifier.height(200.dp).testTag("list"), state = lazyListState) {
           items((0 until 36_000).toList()) { index ->
             BasicText("Item $index")
           }
         }
         UnstyledVerticalScrollbar(
-          scrollAreaState = scrollAreaState,
+          scrollbarState = scrollbarState,
           modifier = Modifier.testTag("scrollbar"),
         ) {
           Thumb(